- DragNDrop now in mostly working condition TODO: fix direct columnService call in the handleOnDragEnd(), to use columnProvider instead
Refactor Board component to use columnService.updatePositions in handleOnDragEnd

The code changes in `Board.jsx` refactor the `handleOnDragEnd` function to use the `columnService.updatePositions` method instead of the `updateUserGameColumnPositions` method. This change ensures that the column positions are updated correctly when dragging and dropping columns on the board. The commented out code related to `setColumnsOnBoard` and `updateUserGameColumnPositions` is removed.