Refactor Board component to use @hello-pangea/dnd library for drag and drop functionality
Leveraged use of React.memo() and useCallbacks to reduce unnecessary functional component rerenders

Used useMemo() to memoize the context value to avoid re-renders  of consuming components

Using SSR to hydrate the userGameProvider and columnProvider with initial data from the DB. Instead of waiting for them to perform a fetch request upon mounting with a useEffect hook.

FIXED: page rerending
the loading states from the context provider  were fullfilling the condition to for the tailloading spinner to render on BoardPageClient. Every time a fetch / update request was made, the load state would become true and always cause the TailSpinner to render instead of the Board component.

TODO: fix search page result rendering the userData with the sear result items .